services/enricher_service/requirements.txt
fastapi
uvicorn[standard]
aiohttp
psycopg2-binary
httpx
pydantic
